backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
sqlalchemy>=2.0.36
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.12
pydantic[email]>=2.10.0
httpx>=0.28.0
python-dotenv>=1.0.1
aiosmtplib>=3.0.0
